% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Adjusted R-Squared Statistic}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{The predicted values from the model}

\item{truth}{The ground truth values}

\item{n_p}{The number of parameters in the model (excluding the intercept)}
}
\description{
Calculates the adjusted R^2 value for a set of estimated and true quantitative values
}
\examples{
pred_r2 = c(5.5, 7.7, 9.9, 11.11, 13.13, 15.15)
truth_r2 = c(5, 8, 12, 4, 18, 20)
adj_R2(pred_r2, truth_r2, 2)
}
