% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{Classes of Tibble Columns}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble}
}
\description{
Returns a character vector containing the classes of each variable in the tibble x.
}
\examples{
test_tibble = as_tibble(data.frame(a = 1:3, b = letters[1:3], c = Sys.Date() - 1:3))
classes(test_tibble)
}
