% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{Numeric Column Scaler}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble}
}
\description{
Scales the numeric columns in a tibble.
}
\examples{
test_tibble = as_tibble(data.frame(a = 1:3, b = letters[1:3], c = Sys.Date() - 1:3))
classes(test_tibble)
}
