% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{Maximum Value Finder}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector of numbers}

\item{na.rm}{Whether to omit NAs or not (default is true)}
}
\value{
A logical vector with TRUE being the maximum value
}
\description{
Accepts an atomic vector x and returns a logical with
TRUE where x equals its maximum value.
}
\examples{
test_vec = c(7, 5, 8, 6)
is_max(test_vec)
}
