% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Chi-Square Log-Likelihood}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{A dataset you want to estimate the likelihood of}

\item{df}{The degrees of freedom of the target chi-square distribution}
}
\description{
Calculates the log likelihood of a given dataset belonging to a chi-squared distribution
}
\examples{
test_chisq = rchisq(1000, 3)
log_likelihood_chisq(test_chisq, 3)
}
