% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{F Log-Likelihood}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A dataset you want to estimate the likelihood of}

\item{df1}{The numerator degrees of freedom of the target F distribution}

\item{df2}{The denominator degrees of freedom of the target F distribution}
}
\description{
Calculates the log likelihood of a given dataset belonging to an F distribution
}
\examples{
test_f = rf(1000, 12, 17)
log_likelihood_f(test_f, 12, 17)
}
