% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Normal Log-Likelihood}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A dataset you want to estimate the likelihood of}

\item{mean}{The mean of the target normal distribution}

\item{sd}{The standard deviation of the target normal distribution}
}
\description{
Calculates the log likelihood of a given dataset belonging to a normal distribution
}
\examples{
test_norm = rnorm(1000, 4, 2.5)
log_likelihood_norm(test_norm, 4, 2.5)
}
