% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{T Log-Likelihood}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A dataset you want to estimate the likelihood of}

\item{df}{The degrees of freedom of the target t distribution}
}
\description{
Calculates the log likelihood of a given dataset belonging to a t distribution
}
\examples{
test_t = rt(1000, 19)
log_likelihood_t(test_t, 19)
}
