% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Uniform Log-Likelihood}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A dataset you want to estimate the likelihood of}

\item{min}{The minimum of the target uniform distribution}

\item{max}{The maximum of the target uniform distribution}
}
\description{
Calculates the log likelihood of a given dataset belonging to a uniform distribution
}
\examples{
test_unif = runif(1000, 0, 1)
log_likelihood_unif(test_unif, 0, 1)
}
