% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Minimum Sample Size from Cohen's d}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{The target effect size}

\item{power}{The desired power of the statistical test}
}
\description{
Calculates the minimum sample size required to perform a two-sample
t-test with a desired Cohen's effect size.
}
\examples{
minimum_n_per_group(0.76)
}
