% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{r2}
\alias{r2}
\title{R-Squared Statistic}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{The estimated values from the model}

\item{truth}{The ground truth values}
}
\description{
Calculates the R^2 value for a set of estimated and true quantitative values
}
\examples{
pred_r2 = c(5.5, 7.7, 9.9, 11.11, 13.13, 15.15)
truth_r2 = c(5, 8, 12, 4, 18, 20)
r2(pred_r2, truth_r2)
}
