% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Sample Wrapper for Vectors and DataFrames}
\usage{
rando(x, n = 1, replace = TRUE)
}
\arguments{
\item{x}{The object you want to sample from}

\item{n}{The number of samples you want to draw}

\item{replace}{Whether to draw samples with or without replacement (default is true)}
}
\value{
The random samples drawn from x
}
\description{
A wrapper around sample that tests whether x is an atomic vector or dataframe-like object
and then returns either n samples or n rows as appropriate
}
\examples{
test_vec = c(5, 6, 7, 8)
rando(test_vec, 4, FALSE)
}
