% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{Matrix Repetition from MATLAB}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A dataframe or matrix to replicate}

\item{M}{How many times to replicate the dataframe or matrix row-wise}

\item{N}{How many times to replicate the dataframe or matrix column-wise}
}
\description{
Accepts a dataframe or matrix and x and returns a matrix created by
replicating the rows (or columns) M (N) times.
}
\examples{
test_df = data.frame(x = c(1, 2, 3, 4), y = c(5, 6, 7, 8))
rep_mat(test_df, 2, 4)
}
